+++ plugins/cleanup_aws_loadbalancers/resoto_plugin_cleanup_aws_loadbalancers/__init__.py
@@ -1,7 +1,7 @@
 from resotolib.baseplugin import BaseActionPlugin
 from resotolib.baseresources import EdgeType
 from resotolib.logging import log
-from resotolib.core.query import CoreGraph
+from resotolib.core.search import CoreGraph
 from resotolib.graph import Graph
 from resoto_plugin_aws.resources import (
     AWSELB,
