+++ plugins/cleanup_aws_vpcs/resoto_plugin_cleanup_aws_vpcs/__init__.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 from resotolib.baseresources import EdgeType
 from resotolib.baseplugin import BaseActionPlugin
 from resotolib.logging import log
-from resotolib.core.query import CoreGraph
+from resotolib.core.search import CoreGraph
 from resotolib.graph import Graph
 from resoto_plugin_aws.resources import (
     AWSVPC,
