+++ plugins/cleanup_untagged/resoto_plugin_cleanup_untagged/__init__.py
@@ -1,7 +1,7 @@
 from copy import deepcopy
 from resotolib.baseplugin import BaseActionPlugin
 from resotolib.logging import log
-from resotolib.core.query import CoreGraph
+from resotolib.core.search import CoreGraph
 from resotolib.core.model_export import node_from_dict
 from resotolib.config import Config
 from .config import CleanupUntaggedConfig
