+++ plugins/protect_snowflakes/resoto_plugin_protect_snowflakes/__init__.py
@@ -1,6 +1,6 @@
 from copy import deepcopy
 from resotolib.logging import log
-from resotolib.core.query import CoreGraph
+from resotolib.core.search import CoreGraph
 from resotolib.baseplugin import BaseActionPlugin
 from resotolib.core.model_export import node_from_dict
 from resotolib.config import Config
