+++ resotolib/resotolib/core/search.py
@@ -43,8 +43,8 @@ class CoreGraph:
             execute_endpoint += f"?{query_string}"
         return self.post(execute_endpoint, command, headers, verify=self.verify)
 
-    def query(self, search: str, edge_type: Optional[EdgeType] = None):
-        log.debug(f"Sending query {search}")
+    def search(self, search: str, edge_type: Optional[EdgeType] = None):
+        log.debug(f"Sending search {search}")
         headers = {"Accept": "application/x-ndjson"}
         search_endpoint = self.search_uri
         if edge_type is not None:
@@ -59,7 +59,7 @@ class CoreGraph:
         r = requests.post(uri, data=data, headers=headers, stream=True, verify=verify)
         if r.status_code != 200:
             log.error(r.content.decode())
-            raise RuntimeError(f"Failed to query graph: {r.content.decode()}")
+            raise RuntimeError(f"Failed to search graph: {r.content.decode()}")
         for line in r.iter_lines():
             if not line:
                 continue
@@ -70,7 +70,7 @@ class CoreGraph:
                 log.error(e)
                 continue
 
-    def graph(self, query: str) -> Graph:
+    def graph(self, search: str) -> Graph:
         def process_data_line(data: dict, graph: Graph):
             """Process a single line of resotocore graph data"""
 
@@ -95,7 +95,7 @@ class CoreGraph:
 
         graph = Graph()
         node_mapping = {}
-        for data in self.query(query):
+        for data in self.search(search):
             try:
                 process_data_line(data, graph)
             except ValueError as e:
